// types.ts
export enum AppStatus {
  IDLE = 'IDLE',
  FILE_SELECTED = 'FILE_SELECTED',
  PROCESSING = 'PROCESSING',
  AWAITING_PASSWORD_FOR_ENCRYPTION = 'AWAITING_PASSWORD_FOR_ENCRYPTION',
  AWAITING_PASSWORD_FOR_DECRYPTION = 'AWAITING_PASSWORD_FOR_DECRYPTION',
  ENCRYPTED = 'ENCRYPTED',
  DECRYPTED = 'DECRYPTED'
}

export interface DecryptedFile {
  url: string;
  fileName: string;
}